+++ packages/icons/src/generated-themed-icons/index.ts
@@ -1,5 +1,6 @@
 export { default as CloudDownloadOutlineMd } from './CloudDownloadOutlineMd';
 export { default as FolderOutlineMd } from './FolderOutlineMd';
 export { default as HomeOutlineMd } from './HomeOutlineMd';
+export { default as LoadingOutlineMd } from './LoadingOutlineMd';
 export { default as SearchOutlineMd } from './SearchOutlineMd';
 export { default as UserOutlineMd } from './UserOutlineMd';