+++ packages/components/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @torr-app/components
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @torr-app/utils@0.0.1
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @torr-app/icons@0.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- Add "Button" component
+
 ## 0.0.2
 
 ### Patch Changes